fixed toggleFollow smell
