+++ src/user/follow.js
@@ -8,64 +8,15 @@ const db = require('../database');
 const utils = require('../utils');
 
 module.exports = function (User) {
+	const followUtils = require('./follow-utils');
+	
 	User.follow = async function (uid, followuid) {
-		await toggleFollow('follow', uid, followuid);
+		return followUtils.handleFollowOperation('follow', uid, followuid, User);
 	};
 
 	User.unfollow = async function (uid, unfollowuid) {
-		await toggleFollow('unfollow', uid, unfollowuid);
+		return followUtils.handleFollowOperation('unfollow', uid, unfollowuid, User);
 	};
-
-	async function toggleFollow(type, uid, theiruid) {
-		if (parseInt(uid, 10) <= 0 || parseInt(theiruid, 10) <= 0) {
-			throw new Error('[[error:invalid-uid]]');
-		}
-
-		if (parseInt(uid, 10) === parseInt(theiruid, 10)) {
-			throw new Error('[[error:you-cant-follow-yourself]]');
-		}
-		const [exists, isFollowing] = await Promise.all([
-			User.exists(theiruid),
-			User.isFollowing(uid, theiruid),
-		]);
-		if (!exists) {
-			throw new Error('[[error:no-user]]');
-		}
-
-		await plugins.hooks.fire('filter:user.toggleFollow', {
-			type,
-			uid,
-			theiruid,
-			isFollowing,
-		});
-
-		if (type === 'follow') {
-			if (isFollowing) {
-				throw new Error('[[error:already-following]]');
-			}
-			const now = Date.now();
-			await db.sortedSetAddBulk([
-				[`following:${uid}`, now, theiruid],
-				[`followers:${theiruid}`, now, uid],
-			]);
-		} else {
-			if (!isFollowing) {
-				throw new Error('[[error:not-following]]');
-			}
-			await db.sortedSetRemoveBulk([
-				[`following:${uid}`, theiruid],
-				[`followers:${theiruid}`, uid],
-			]);
-		}
-
-		const [followingCount, followingRemoteCount, followerCount, followerRemoteCount] = await db.sortedSetsCard([
-			`following:${uid}`, `followingRemote:${uid}`, `followers:${theiruid}`, `followersRemote:${theiruid}`,
-		]);
-		await Promise.all([
-			User.setUserField(uid, 'followingCount', followingCount + followingRemoteCount),
-			User.setUserField(theiruid, 'followerCount', followerCount + followerRemoteCount),
-		]);
-	}
 
 	User.getFollowing = async function (uid, start, stop) {
 		return await getFollow(uid, 'following', start, stop);
